requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
psycopg2-binary>=2.9.0
flask>=3.0.0
streamlit>=1.30.0
confluent-kafka>=2.3.0
